Refactor source and add doc comments
Also refines agent instructions.
