+++ .cspell.json
@@ -3,7 +3,8 @@
   "version": "0.2",
   "useGitignore": true,
   "ignorePaths": [
-    ".vscode/"
+    ".vscode/",
+    "tests/**/Content/"
   ],
   "words": [
     "accessibilities",
