+++ src/AzureSdk.SamplesMcp/AssemblyInfo.cs
@@ -1,3 +1,6 @@
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("AzureSdk.SamplesMcp.Test")]
