+++ src/AzureSdk.SamplesMcp/Program.cs
@@ -1,4 +1,7 @@
-﻿using AzureSdk.SamplesMcp;
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
+using AzureSdk.SamplesMcp;
 using AzureSdk.SamplesMcp.Services;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
