+++ src/AzureSdk.SamplesMcp/Services/Environment.cs
@@ -1,4 +1,7 @@
-namespace AzureSdk.SamplesMcp;
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
+namespace AzureSdk.SamplesMcp.Services;
 
 // cspell:words USERPROFILE
