+++ src/AzureSdk.SamplesMcp/Services/JsonPath.cs
@@ -1,3 +1,6 @@
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
 using System.Text.Json;
 
 namespace AzureSdk.SamplesMcp.Services;
