+++ src/AzureSdk.SamplesMcp/Services/ServicesJsonContext.cs
@@ -1,10 +1,13 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
 using System.Text.Json.Serialization;
 
 namespace AzureSdk.SamplesMcp.Services;
 
+/// <summary>
+/// Source-generated JSON context for service-level payloads.
+/// </summary>
 [JsonSerializable(typeof(ExternalProcessService.ParseError))]
 [JsonSerializable(typeof(ExternalProcessService.ParseOutput))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
