+++ src/AzureSdk.SamplesMcp/Tools.cs
@@ -1,3 +1,6 @@
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
 using System.ComponentModel;
 using AzureSdk.SamplesMcp.Providers;
 using AzureSdk.SamplesMcp.Services;
