+++ tests/AzureSdk.SamplesMcp.Test/MSTestSettings.cs
@@ -1,1 +1,4 @@
-﻿[assembly: Parallelize(Scope = ExecutionScope.MethodLevel)]
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
+[assembly: Parallelize(Scope = ExecutionScope.MethodLevel)]
