+++ tests/AzureSdk.SamplesMcp.Test/Providers/NodeTests.cs
@@ -1,10 +1,11 @@
-using System.Reflection;
-using AzureSdk.SamplesMcp.Providers;
-using AzureSdk.SamplesMcp.Test.Services;
-using Microsoft.Extensions.FileProviders;
-using Microsoft.Extensions.Logging;
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-namespace AzureSdk.SamplesMcp;
+using System.Reflection;
+using AzureSdk.SamplesMcp.Services;
+using Microsoft.Extensions.FileProviders;
+
+namespace AzureSdk.SamplesMcp.Providers;
 
 [TestClass]
 public class NodeTests
