+++ tests/AzureSdk.SamplesMcp.Test/Services/FileSystemTests.cs
@@ -1,7 +1,10 @@
+// Copyright 2026 Heath Stewart.
+// Licensed under the MIT License. See LICENSE.txt in the project root for license information.
+
 using System.Reflection;
 using Microsoft.Extensions.FileProviders;
 
-namespace AzureSdk.SamplesMcp;
+namespace AzureSdk.SamplesMcp.Services;
 
 [TestClass]
 public class FileSystemTests
